vin_ocr/requirements.txt
# PaddleOCR VIN Recognition Pipeline - Unified Requirements
# ===========================================================
# Python 3.11+ recommended
# Environment: paddleocr_vin
#
# Installation:
#   1. Create environment: conda create -n paddleocr_vin python=3.11
#   2. Activate: conda activate paddleocr_vin
#   3. Install: pip install -r requirements.txt
#
# For GPU support, see installation instructions in INSTALLATION.md

# ============================================
# Core Deep Learning Frameworks
# ============================================

# Pipeline orchestration / tracking
zenml==0.92.0
mlflow==3.7.0
dvc==3.64.2

# PaddlePaddle (choose ONE based on your hardware)
# For GPU (CUDA 12.x) - RECOMMENDED for training:
# paddlepaddle-gpu==3.0.0
# For CPU only (inference/testing):
paddlepaddle==3.0.0

# OCR Engine
paddleocr==3.3.3

# ============================================
# Computer Vision & Image Processing
# ============================================
opencv-python==4.13.0.90
opencv-contrib-python==4.10.0.84
Pillow==12.1.0
scikit-image==0.26.0
imageio==2.37.2

# ============================================
# Deep Learning & ML Utilities
# ============================================
numpy==2.4.1
scipy==1.17.0
scikit-learn==1.5.2

# Data augmentation
albumentations==2.0.8
albucore==0.0.24

# ============================================
# Optional: PyTorch & Transformers (for DeepSeek-OCR / LiquidAI)
# ============================================
# DeepSeek-OCR requires specific versions - tested on Python 3.12.9 + CUDA 11.8
# For CUDA wheels:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For Flash Attention (NVIDIA GPU only, required for DeepSeek-OCR):
#   pip install flash-attn==2.7.3 --no-build-isolation
# For non-CUDA systems, install the stub instead:
#   pip install -e ./stubs/flash_attn_stub
torch==2.6.0
torchvision==0.21.0
transformers==4.46.3
tokenizers==0.20.3
accelerate==1.12.0
peft==0.18.1
timm==1.0.24
einops==0.8.0
datasets==3.0.1

# DeepSeek-OCR specific dependencies
addict>=2.4.0
easydict>=1.10

# LiquidAI requires Unsloth for loading lfm2_vl architecture
# See: https://github.com/Liquid4All/cookbook
# NOTE: Unsloth requires NVIDIA GPU (CUDA) - does NOT work on Apple Silicon/CPU
# For GPU systems only: pip install unsloth triton
# unsloth  # Uncomment on CUDA systems to enable LiquidAI model support
trl>=0.9.0  # Required by Unsloth for SFTTrainer

# Optional GPU optimization
# bitsandbytes>=0.41.0
# deepspeed>=0.12.0
# onnx>=1.15.0
# onnxruntime-gpu>=1.16.0

# ============================================
# Configuration & Data Management
# ============================================
pyyaml==6.0.2
pydantic==2.11.9
pydantic-settings==2.12.0

# ============================================
# Data Processing & Analysis
# ============================================
pandas==2.3.3
rapidfuzz==3.14.3

# ============================================
# Web Interface (Streamlit)
# ============================================
streamlit==1.53.1
plotly==6.5.2
altair==6.0.0

# ============================================
# Utilities & Tools
# ============================================
tqdm==4.67.1
requests==2.32.5
click==8.2.1
fsspec==2024.10.0
s3fs==2024.10.0

# Text processing
python-bidi==0.6.7
shapely==2.1.2
pyclipper==1.4.0

# File operations
lmdb==1.7.5
pypdfium2==5.3.0

# ============================================
# Testing & Development
# ============================================
pytest==9.0.2

# Optional: Browser testing for web UI
# playwright==1.49.0
# pytest-playwright==0.6.2

# ============================================
# Notebook Support (Optional)
# ============================================
# ipykernel==6.29.5
# jupyter-client==8.6.3

# ============================================
# Model Training & Monitoring (Optional)
# ============================================
# For TensorBoard/VisualDL logging
# visualdl>=2.5.0
# tensorboard>=2.10.0

# Hyperparameter tuning
optuna>=4.0.0

# ============================================
# Version Management Notes
# ============================================
# Compatible with:
# - Python: 3.11+
# - CUDA: 12.4+ (for GPU version)
# - Windows 10/11, Linux, macOS
#
# Last updated: January 2026
# For issues, see: INSTALLATION.md

# Linting/Formatting (development only)
# ruff==0.6.9
# black==24.10.0

# =============================================================================
# NEURAL NETWORK FINE-TUNING (GPU RECOMMENDED)
# =============================================================================
# For full neural network training with 11,000+ images, use GPU version:
#   pip uninstall paddlepaddle
#   pip install paddlepaddle-gpu
#
# For multi-GPU training:
#   python -m paddle.distributed.launch --gpus '0,1,2,3' finetune_paddleocr.py
#
# For PaddleOCR advanced training features, clone their repo:
#   git clone https://github.com/PaddlePaddle/PaddleOCR.git
#   cd PaddleOCR && pip install -r requirements.txt

# =============================================================================
# HPC REQUIREMENTS (RTX 3090 24GB - DeepSeek Fine-Tuning)
# =============================================================================
# For DeepSeek-OCR fine-tuning on HPC with NVIDIA RTX 3090:
#
# Core requirements:
#   torch>=2.0.0+cu118  # PyTorch with CUDA 11.8
#   transformers>=4.36.0
#   accelerate>=0.25.0
#   peft>=0.6.0  # For LoRA fine-tuning
#
# Memory optimization (optional, for large models):
#   bitsandbytes>=0.41.0  # 8-bit/4-bit quantization
#   deepspeed>=0.12.0  # Distributed training
#
# ONNX export:
#   onnx>=1.15.0
#   onnxruntime-gpu>=1.16.0  # For GPU inference
#   # Or: onnxruntime>=1.16.0  # For CPU-only inference
#
# RTX 3090 (24GB) Recommended Settings:
#   - Use LoRA with bf16: batch_size=4, gradient_accumulation=8
#   - Enable gradient_checkpointing for larger models
#   - Use 8-bit quantization if running out of memory
#
# Installation on HPC:
#   module load cuda/11.8
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#   pip install transformers accelerate peft bitsandbytes
#   pip install onnx onnxruntime-gpu
